Fix 2Byte fixed SE-HUFFMAN encoding error

+++ MSDB/MSDB-CORE/include/compression/adaptHuffmanChunk.h
@@ -46,7 +46,7 @@ private:
 
 		//////////////////////////////
 		// Adaptive huffman coding
-		aHuffmanCoder<uint32_t, uint8_t> coder(sizeof(uint8_t) * CHAR_BIT);
+		aHuffmanCoder<uint16_t, uint8_t> coder(sizeof(uint8_t) * CHAR_BIT);
 		auto buffer = (const char*)this->getBuffer()->getReadData();
 		auto size = this->getBuffer()->size();
 		for (int i = 0; i < size; ++i)
